refactor: Restructure project with src directory and update imports
- Move main project files into `src` directory
- Update Dockerfile to use new main.py path
- Modify pyproject.toml to include `src` package
- Update import statements in tests and main files
- Add `__init__.py` files to new package structure
