+++ src/actions/code_scan.py
@@ -3,7 +3,7 @@ import os
 import logging
 from typing import Any
 
-from github_agent import GitHubAgent
+from src.github_agent import GitHubAgent
 
 logger = logging.getLogger("code-scan-action")
